vim-patch:8.1.0825: code for autocommands is mixed with file I/O code
