+++ src/nvim/fileio.h
@@ -3,6 +3,7 @@
 
 #include "nvim/buffer_defs.h"
 #include "nvim/os/os.h"
+#include "nvim/autocmd.h"
 
 // Values for readfile() flags
 #define READ_NEW        0x01    // read a file into a new buffer
@@ -15,23 +16,8 @@
 
 #define READ_STRING(x, y) (char_u *)read_string((x), (size_t)(y))
 
-/*
- * Struct to save values in before executing autocommands for a buffer that is
- * not the current buffer.
- */
-typedef struct {
-  buf_T       *save_curbuf;     ///< saved curbuf
-  int use_aucmd_win;            ///< using aucmd_win
-  win_T       *save_curwin;     ///< saved curwin
-  win_T       *save_prevwin;    ///< saved prevwin
-  win_T       *new_curwin;      ///< new curwin
-  bufref_T     new_curbuf;      ///< new curbuf
-  char_u      *globaldir;       ///< saved value of globaldir
-} aco_save_T;
-
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 // Events for autocommands
-# include "auevents_enum.generated.h"
 # include "fileio.h.generated.h"
 #endif
 #endif  // NVIM_FILEIO_H
